some code cleaning

+++ R/divime_classify_vox.R
@@ -26,7 +26,7 @@ divime_classify_vox <- function(audio_loc,
   vm_running <- divime_vagrant_state(divime_loc = divime_loc,
                                      what = "status",
                                      silent = TRUE)
-  if(!vm_running) {
+  if (!vm_running) {
     divime_vagrant_state(divime_loc = divime_loc,
                          what = "start",
                          silent = TRUE)
@@ -35,10 +35,10 @@ divime_classify_vox <- function(audio_loc,
   paths <- avutils:::handle_filenames(audio_loc = audio_loc,
                                       divime_loc = divime_loc)
 
-  # set command
-  cm <- paste0("ssh -c 'vcm.sh data/'")
-
   logres <- data.frame(audio = paths$filestoprocess,
+                       size = paths$size,
+                       processed = FALSE,
+                       ptime = NA,
                        yunifile = NA,
                        yunicopy = NA,
                        yuniremove = NA,
@@ -48,52 +48,92 @@ divime_classify_vox <- function(audio_loc,
                        resultscopy = NA,
                        resultsremove = NA) #, yuniproblem = NA
 
+  # set command
+  cm <- paste0("ssh -c 'vcm.sh data/'")
+
   # loop through files
   for (i in 1:nrow(logres)) {
-    # copy audio
-    logres$audiocopy[i] <- file.copy(from = paths$audiosource[i],
-                                     to = paths$audiotarget_clean[i])
+    # take time stamp
+    t1 <- Sys.time()
 
     # only run if the yuni source was found in the source folder...
     yunifrom <- paste0(audio_loc, "/", paths$folder[i], "yunitator_old_", paths$root[i], ".rttm")
     yunito <- paste0(divime_loc, "/data/", paste0("yunitator_universal_", paths$root_clean[i], ".rttm"))
 
     if (file.exists(yunifrom)) {
-      logres$yunifile[i] <- paste0("yunitator_old_", paths$root[i], ".rttm")
-      logres$yunicopy[i] <- file.copy(from = yunifrom,
-                                     to = yunito)
-
-      WD <- getwd()
-      setwd(divime_loc)
-      xres <- system2(command = vagrant, args = cm, stdout = TRUE, stderr = TRUE)
-      setwd(WD)
-
-      logres$audioremove[i] <- file.remove(paths$audiotarget_clean[i])
-      logres$yuniremove[i] <- file.remove(yunito)
-
       output_file <- paste0("vcm_", paths$root_clean[i], ".rttm")
       output_file_ori <- paste0("vcm_", paths$root[i], ".rttm")
+      output_file_to <- normalizePath(paste0(audio_loc, "/", paths$folder[i], output_file_ori),
+                                      winslash = "/",
+                                      mustWork = FALSE)
+      output_file_from <- normalizePath(paste0(divime_loc, "/data/", output_file),
+                                        winslash = "/",
+                                        mustWork = FALSE)
 
-      outpath <- paste0(audio_loc, "/", paths$folder[i], output_file_ori)
+      # if overwrite = FALSE, continue only if the target file does not yet exist
+      # if it already exists, we can skip the processing in the VM
+      output_exists <- file.exists(output_file_to)
+      if (!(!overwrite & output_exists)) {
+        # copy audio
+        logres$audiocopy[i] <- file.copy(from = paths$audiosource[i],
+                                         to = paths$audiotarget_clean[i])
+        logres$yunifile[i] <- paste0("yunitator_old_", paths$root[i], ".rttm")
+        logres$yunicopy[i] <- file.copy(from = yunifrom,
+                                        to = yunito)
+        # deal with working directories
+        WD <- getwd()
+        setwd(divime_loc)
 
-      logres$resultscopy[i] <- file.copy(from = paste0(divime_loc, "/data/", output_file),
-                                         to = suppressWarnings(outpath),
-                                         overwrite = overwrite)
-      logres$resultsremove[i] <- file.remove(paste0(divime_loc, "/data/", output_file))
+        # run bash command
+        xres <- system2(command = vagrant, args = cm, stdout = TRUE, stderr = TRUE)
+        setwd(WD)
 
-      logres$output[i] <- output_file_ori
+        # remove source files (audio and ynui)
+        logres$audioremove[i] <- file.remove(paths$audiotarget_clean[i])
+        logres$yuniremove[i] <- file.remove(yunito)
 
-      if (messages) message(paths$filestoprocess[i], "  -->  ", output_file_ori)
+        # copy output back to source location from divime location
+        logres$resultscopy[i] <- file.copy(from = output_file_from,
+                                           to = output_file_to,
+                                           overwrite = overwrite)
+        logres$resultsremove[i] <- file.remove(output_file_from)
+
+        logres$output[i] <- output_file_ori
+        logres$processed[i] <- TRUE
+
+        if (messages) message(paths$filestoprocess[i], "  -->  ", output_file_ori)
+
+        # clean up
+        rm(xres, WD)
+      } else {
+        if (messages) message(paths$filestoprocess[i], "  -->  ", "XXXXXXXXX")
+      }
 
       # clean up
-      rm(outpath, output_file, output_file_ori, xres)
+      rm(output_exists, output_file, output_file_from, output_file_ori, output_file_to)
 
     } else {
       if (messages) message(paths$filestoprocess[i], "  -->  ", "XXXXXXXXX")
+      logres$yunifile[i] <- "not found"
     }
 
     # more clean up
     rm(yunifrom, yunito)
+
+    # time stamp again
+    t2 <- Sys.time()
+    logres$ptime[i] <- as.numeric(round(difftime(t2, t1, unit = "min"), 3))
+
+    # predict time left
+    temp <- na.omit(logres[, c("ptime", "size")])
+    sizes <- logres$size[is.na(logres$ptime)]
+    if (nrow(temp) > 3) {
+      tempres <- lm(ptime ~ size, temp)
+      if (length(sizes) > 0) {
+        timeleft <- round(sum(predict(tempres, newdata = data.frame(size = sizes))), 1)
+        cat("expected time until finish: ", timeleft, " minutes\n")
+      }
+    }
   }
 
   # shut down if requested
